+++ src/UBSAg.Negotiations.Application/Responses/TradesQueryResponse.cs
@@ -3,7 +3,7 @@ namespace UBSAg.Negotiations.Application.Responses
 {
     public class TradesQueryResponse
     {
-        public decimal Value { get; set; }
+        public double Value { get; set; }
         public string ClientSector { get; set; }
         public string? TradeCategory { get; set; }
     }
